Reorganize project structure and begin cleaning
