+++ src/main/java/Main.java
@@ -1,5 +1,6 @@
 import javax.swing.*;
-import SQL.CreateSQLTables;
+
+import GUI.DrinkGUI;
 
 
 public class Main{
@@ -7,15 +8,11 @@ public class Main{
 	
 	
 	public static void main(String[] args)throws ClassNotFoundException{
-		
-		
 		SwingUtilities.invokeLater(new Runnable(){
 			@Override
 			public void run() {
 			      	new DrinkGUI().setVisible(true);
 			}
 		});
-		//Creates drinks table in MySQL DB
-		//CreateSQLTables drinkTable = new CreateSQLTables();
 	}
 }
